Move Android deploy dialog out of settings page
