+++ flutter/lib/main.dart
@@ -27,7 +27,7 @@ import 'common.dart';
 import 'consts.dart';
 import 'mobile/pages/home_page.dart';
 import 'mobile/pages/server_page.dart';
-import 'mobile/pages/settings_page.dart';
+import 'mobile/widgets/deploy_dialog.dart';
 import 'models/platform_model.dart';
 
 import 'package:flutter_hbb/plugin/handlers.dart'
@@ -577,8 +577,8 @@ _registerEventHandler() {
     });
   }
   if (isAndroid) {
-    platformFFI.registerEventHandler('android_needs_deploy',
-        'android_needs_deploy', (_) async {
+    platformFFI.registerEventHandler(
+        'android_needs_deploy', 'android_needs_deploy', (_) async {
       WidgetsBinding.instance.addPostFrameCallback((_) {
         showDeployPromptDialog();
       });
